semana 6 actualizado

+++ src/main/java/com/tiendat/domain/Categoria.java
@@ -16,7 +16,7 @@ public class Categoria implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_categoria")
-    private Long idCategoria;
+    private Integer idCategoria;
     @NotBlank(message = "La descripción no puede estar vacía")
     @Size(max = 50, message = "La descripción no puede exceder los 50 caracteres")
     @Column(length = 50, nullable = false, unique = true)
