+++ src/main/java/com/tiendat/service/FirebaseStorageService.java
@@ -20,8 +20,7 @@ public class FirebaseStorageService {
 
     @Value("${firebase.storage.path}")
     private String storagePath;
-
-    // Cliente de Firebase Storage (inyectado)
+    // Aquí se manejaría la inyección del cliente de Storage como un bean
     private final Storage storage;
 
     // 🔧 Constructor corregido (nombre igual a la clase)
@@ -29,8 +28,8 @@ public class FirebaseStorageService {
         this.storage = storage;
     }
 
-    // 📤 Sube un archivo de imagen al almacenamiento de Firebase
-    public String uploadImage(MultipartFile localFile, String folder, Long id) throws IOException {
+    //Sube un archivo de imagen al almacenamiento de Firebase.    
+    public String uploadImage(MultipartFile localFile, String folder, Integer id) throws IOException {
         String originalName = localFile.getOriginalFilename();
         String fileExtension = "";
 
@@ -38,22 +37,23 @@ public class FirebaseStorageService {
             fileExtension = originalName.substring(originalName.lastIndexOf("."));
         }
 
-        // Nombre del archivo con formato consistente
+        // Se genera el nombre del archivo con un formato consistente.
         String fileName = "img" + getFormattedNumber(id) + fileExtension;
 
         File tempFile = convertToFile(localFile);
+
         try {
             return uploadToFirebase(tempFile, folder, fileName);
         } finally {
-            // Elimina el archivo temporal
+            // Asegura que el archivo temporal se elimine siempre.
             if (tempFile.exists()) {
                 tempFile.delete();
             }
         }
     }
 
-    // 🧩 Convierte un MultipartFile a un archivo temporal
-    private File convertToFile(MultipartFile multipartFile) throws IOException {
+    //Convierte un MultipartFile a un archivo temporal en el servidor.
+     private File convertToFile(MultipartFile multipartFile) throws IOException {
         File tempFile = File.createTempFile("upload-", ".tmp");
         try (FileOutputStream fos = new FileOutputStream(tempFile)) {
             fos.write(multipartFile.getBytes());
@@ -61,22 +61,25 @@ public class FirebaseStorageService {
         return tempFile;
     }
 
-    // ☁️ Sube el archivo al almacenamiento de Firebase y genera una URL firmada
+    //Sube el archivo al almacenamiento de Firebase y genera una URL firmada.     
     private String uploadToFirebase(File file, String folder, String fileName) throws IOException {
+        // Definimos el ID del blob y su información
         BlobId blobId = BlobId.of(bucketName, storagePath + "/" + folder + "/" + fileName);
         String mimeType = Files.probeContentType(file.toPath());
-        BlobInfo blobInfo = BlobInfo.newBuilder(blobId)
-                .setContentType(mimeType != null ? mimeType : "media")
-                .build();
+        BlobInfo blobInfo = BlobInfo.newBuilder(blobId).setContentType(mimeType != null ? mimeType : "media").build();
 
-        // Subimos el archivo
+        // Subimos el archivo. El objeto `storage` ya tiene las credenciales necesarias.
         storage.create(blobInfo, Files.readAllBytes(file.toPath()));
 
-        // URL firmada válida por 5 años
+        // El objeto `storage` ya tiene las credenciales del servicio configuradas        
+        // Se genera la URL firmada. Ahora con una caducidad de 5 años.
         return storage.signUrl(blobInfo, 1825, TimeUnit.DAYS).toString();
     }
 
-    // 🔢 Formatea el ID en un número de 14 dígitos con ceros a la izquierda
+    /**
+     * Genera un string numérico con un formato de 14 dígitos, rellenado con
+     * ceros a la izquierda.
+     */
     private String getFormattedNumber(long id) {
         return String.format("%014d", id);
     }
